Move PollList to Pages folder
-Change text color for account name
-Move Poll List to pages folder
